pyproject.toml
[project]
name = "noisy-agentdojo"
version = "0.1.0"
description = "Noise-augmented adversarial training data pipeline built on AgentDojo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agentdojo>=0.1.35",
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "icalendar>=5.0",
    "vobject>=0.9",
    "tenacity>=8.2",
    "rich>=13.0",
    "google-genai>=1.0",
]

[project.scripts]
noisy-agentdojo = "noisy_agentdojo.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
